Going very slowly
I think I am too far out of the target audience for
this person's videos. He is moving agonizingly slowly
for me, repeating concepts I learned decades ago over
and over and over again, with too little code for
demonstration. The code itself, once he gets around to
it is fine.

This is not a criticism of the instructor as a person.
He seems like very nice person. This is simply an expression of my
frustration at the pace of the course.

One technical thing: he set up an automatic watch and compile
command many videos ago, and for whatever reason it is no longer
working for me...which is my main frustration with Javascript:
the goddamn tooling is miserable.