video 84, updste algo
This video series is just agoonizingly slow. Even the simplest
things are explained three times or more, sometimes in the same
way. With everything else I have going on right now, it's all I
can do to just a commitable piece of code out of a video and move
on. This is stretching out the time it takes for the videos, which
makes the problem even worse.

This is not the fault of the author. I am just very far out of his
intended audience.