stats - start of project to ingest csv
I took a break from this project for a while, which was good, I
needed to get away from it for a while.

This commit represents project set up. I took some notes on in
my paper notebook as it's always difficult for me to recall
exactly how these javascript projects are set up.