240 proof of concept
This isn't working due to a problem with the node/typescript
system.

The whole course reflects what I consider a fundamental problem
in the entire Javascript community: the default testing framework
is Looks Right. That is, write some code, refresh the browser to
see what happens. This sort of testing makes it really difficult
to pin down errors. I would have to manually bisect the last
many commits, which is really hard as there is no "looks right"
defined for the last commits.

I realize this may well be a mistaken impression, but it's my
impression nonetheless: the Javascript community remains immature
with respect to solid engineering.

My impression has been formed as a relatively experienced engineer
in other technologies who is self-teaching Javascript. The
experience is markedly different than when I self-taught Ruby
and Rails 10 years ago. Testing was front and center.