257 - simple components
This one is broken right out of the box, the react
typescript components apparently are not correctly
exported from the type definition file, or something.
Not a good sign this early into this module.

Once again, the javascript ecosystem does not fail to
disappoint. Shit is broken right out of the box.